remove the encoder (use raw inputs as memory)
having an RNN encoder that generates memory units and training it meant
that the pattern of information distribution over memory units changed
over time. This resulted in the attention becoming more screwed with
more training (probably due to the network trying to bring more
information from far away places into the first few memory units, and
thereby making the attention be fixated on those first few units even
when generating outputs towards the end).

Having a uniform distribution of information in the memory units is
necessary for attention to work best, I believe.